Config.validate_webhook -> Pusher.validate_webhook
