+++ pusher_tests/test_config.py
@@ -99,63 +99,5 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(conf.host, u'host')
         self.assertEqual(conf.app_id, u'4')
 
-    def test_validate_webhook_success_case(self):
-        conf = Config.from_url(u'http://foo:bar@host/apps/4')
-
-        body = u'{"time_ms": 1000000}'
-        signature = six.text_type(hmac.new(conf.secret.encode(u'utf8'), body.encode(u'utf8'), hashlib.sha256).hexdigest())
-
-        with mock.patch('time.time', return_value=1200):
-            self.assertEqual(conf.validate_webhook(conf.key, signature, body), {u'time_ms': 1000000})
-
-    def test_validate_webhook_bad_types(self):
-        conf = Config.from_url(u'http://foo:bar@host/apps/4')
-
-        conf.validate_webhook(u'key', u'signature', u'body')
-
-        # These things are meant to be human readable, so enforcing being text is
-        # sensible.
-
-        with mock.patch('time.time') as time_mock:
-            self.assertRaises(TypeError, lambda: conf.validate_webhook(4, u'signature', u'body'))
-            self.assertRaises(TypeError, lambda: conf.validate_webhook(b'test', u'signature', u'body'))
-            self.assertRaises(TypeError, lambda: conf.validate_webhook(u'key', 4, u'body'))
-            self.assertRaises(TypeError, lambda: conf.validate_webhook(u'key', b'signature', u'body'))
-            self.assertRaises(TypeError, lambda: conf.validate_webhook(u'key', u'signature', 4))
-            self.assertRaises(TypeError, lambda: conf.validate_webhook(u'key', u'signature', b'body'))
-
-        time_mock.assert_not_called()
-
-    def test_validate_webhook_bad_key(self):
-        conf = Config.from_url(u'http://foo:bar@host/apps/4')
-
-        body = u'some body'
-        signature = six.text_type(hmac.new(conf.secret.encode(u'utf8'), body.encode(u'utf8'), hashlib.sha256).hexdigest())
-
-        with mock.patch('time.time') as time_mock:
-            self.assertEqual(conf.validate_webhook(u'badkey', signature, body), None)
-
-        time_mock.assert_not_called()
-
-    def test_validate_webhook_bad_signature(self):
-        conf = Config.from_url(u'http://foo:bar@host/apps/4')
-
-        body = u'some body'
-        signature = u'some signature'
-
-        with mock.patch('time.time') as time_mock:
-            self.assertEqual(conf.validate_webhook(conf.key, signature, body), None)
-
-        time_mock.assert_not_called()
-
-    def test_validate_webhook_bad_time(self):
-        conf = Config.from_url(u'http://foo:bar@host/apps/4')
-
-        body = u'{"time_ms": 1000000}'
-        signature = six.text_type(hmac.new(conf.secret.encode('utf8'), body.encode('utf8'), hashlib.sha256).hexdigest())
-
-        with mock.patch('time.time', return_value=1301):
-            self.assertEqual(conf.validate_webhook(conf.key, signature, body), None)
-
 if __name__ == '__main__':
     unittest.main()
